Fixed NumberSum solution

+++ Praktikum 05/Latihan/B/NumberSum.java
@@ -4,37 +4,30 @@ public class NumberSum {
         int idx = 0;
 
         for (int i = 0; i < numArr.length; i++) {
-            // System.out.println(i);
             if (numArr[i] != null) {
                 ans[idx] = numArr[i];
-                // System.out.println(ans[idx]);
                 for (int j = i + 1; j < numArr.length; j++) {
-                    // System.out.println(ans[idx] + ", " + numArr[j]);
                     if (numArr[j] == null) continue;
                     if (numArr[i].getClass().equals(numArr[j].getClass())) {
                         if (numArr[j] instanceof Byte) {
-                            ans[idx] = (Byte) ans[idx] + (Byte) numArr[j];
+                            ans[idx] = (byte) (ans[idx].byteValue() + numArr[j].byteValue());
                         } else if (numArr[j] instanceof Integer) {
-                            ans[idx] = (Integer) ans[idx] + (Integer) numArr[j];
+                            ans[idx] = (int) (ans[idx].intValue() + numArr[j].intValue());
                         } else if (numArr[j] instanceof Double) {
-                            ans[idx] = (Double) ans[idx] + (Double) numArr[j];
+                            ans[idx] = (double) (ans[idx].doubleValue() + numArr[j].doubleValue());
                         } else if (numArr[j] instanceof Float) {
-                            ans[idx] = (Float) ans[idx] + (Float) numArr[j];
+                            ans[idx] = (float) (ans[idx].floatValue() + numArr[j].floatValue());
                         } else if (numArr[j] instanceof Short) {
-                            ans[idx] = (Short) ans[idx] + (Short) numArr[j];
+                            ans[idx] = (short) (ans[idx].shortValue() + numArr[j].shortValue());
                         } else if (numArr[j] instanceof Long) {
-                            ans[idx] = (Long) ans[idx] + (Long) numArr[j];
-                        }
+                            ans[idx] = (long) (ans[idx].longValue()+ numArr[j].longValue());
+                        } 
                         numArr[j] = null;
                     }
                 }
                 idx++;
             }
-
-            // System.out.println(ans[idx]);
-
         }
-
         return ans;
     }
 }
